add fallback tag to mirror fallback-mount.
while the duplication seems unnecessary, the idea is to utilise the newer xml
work while maintaining compatability and to help with readability.

Internally the fallback mount is now encompassed into the fbinfo structure
which also maintain the target rate in the case of files. So you can now use
the traditional

<fallback-mount>/something</fallback-mount>
or
<fallback>/something</fallback>
<fallback mount="/something" />
<fallback> <mount>/something</mount> </fallback>

or add extra like

<fallback mount="/something" rate="128k" />
or some combo style like above.

Most changes were realting to referencing the name and rate, and I changed a
member name to avoid a double barrel.  The override and fallback on full
settings stay in the mount for now as they really state a policy for mount.