config parsing update
The basic handling is the same but differences in the internal API for handling
conditions.  Extra flags for handling processing.

There is a helper function added to wrap around the retrieve callback. This is
mainly for preparing the data for parsing depending on whether it is a property
or a node.

Updates for the config_get_* routines, mostly to improve checks. We should now
be ok for oddball cases as well like empty tags not failing.

Warnings are reported for more things like duplication when not expected. They
still apply but you are notified of it.