reduce duplicate http header logging on reload
while logging the lines on reload from the xml once is fine, for every mount
could be wasteful as if would be exactly the same for each one. It would also
be before any changes and you do not seen the mount name as the lines for the
mount are done as a default before the xml is parsed so you don't know the mount
name when they are copied.  So only log the new entries and not copied ones.