do relay start check in relay_initialise to avoid source setup
the _initialise routine sets the initial source parameters like on demand and
the stats. This means that a relay that is in a delayed restart will have the
stats appear and it will allow listeners to connect onto it, causing them to
pause until the relay actually starts.  Here we add the start time check into
the init routine before the on-demand relay setup.