when starting with a fallback file, the initialization was incomplete
for icy streams when starting with a fallback file, the metadata setting was
not setup because the mount information had not been applied like it was for
streams.  The actual setup occurs via the query count path but open_fh did not
have the mount info to apply, so we needed to pass that down.

karl.