minor xslt update
expand on the xsl cache size, unlikely to be an issue for anyone now.

I'll keep the xslt lock for a little bit longer to make sure of caching. We
should drop the client staying on the worker while ACTIVE status is off as it
still requires locking as the worker still reads the flags.  For now, just
expand the range of this lock.