fix for stale memory use case
the stream start trigger uses a dummy client to prevent the object disappearing
on a terminating source, so we need to populate it minimally. For url auth, we
use the aux_data reference for any intro parsing so we should not use that for
an agent copy.

If the number of items increases then we may need to provide either a struct or
separate callbacks for treatin differently

The flag change is minor operationally but prevents a warning message.