active status changes.
This changes the way clients are treated if they are being processed by separate
threads.  Instead of unsetting the ACTIVE flag (making the worker skip it), we
trigger the client to be removed from the worker using the balancing trigger.

The worker then ignores this client and drops from the list but here we do what
ever processing is needed then re-add to the incoming worker.

This simplifies the coding as we really need to lock the worker to set the
ACTIVE status and schedule when re-adding which makes things ugly. Here the
re-add routine nicely covers that.

Key aspect to note is the worker needs to be nulld before you leaving client or
else it could end up ina confused state.