update to xslt client handling.
this extends on the previous work of hiving off xsl processing off the workers
into their own thread. Also when rejoing workers, it put them on the incoming
worker to limit processing for listenrs.  This though deals mainly with the
xsl cache interactions.

The cache is not large as there really is not that many xsl files but this cache
is also keeping failed entries flagged up to reduce lookups in such cases. All
translations are still performed via isolated threads as the libs can be quite
involved.  failed slots are selected before successful ones and retries are
still possible but seems harder now to trigger.