yp routines update
Basic operation is the same but avoid reduce some memory use.

The ypdata was storing copies of the stream metadata escape encoded ready for
the POST.  I'm now using the param routines to build the POST for add and touch
handlers (the remove only uses the action and sid which are not encoded).

As the param code can now handle escape encoding, then much of the pre-encoded
stream meta is no longer needed and is now filtered out.  The server id is just
copied as needed and admin is added the POST for add.  yp_stat is just a wrapper
for stats_retrieve_nocopy but return empty string instead of null.