fix possible data race triggered from stats handling
possible race between the stats created for non-present sources with fallbacks
and relays that start up.  We need to keep the source tree lock while we
determine if stats need to be provided.  A source may be present with delayed
relays but not fully restarted, but sets up stats to show its presence.

This now looks cleaner than before