top autoconf/make changes to improve build
the following helps in doing out of tree builds which helps keep things clearer
The new .in files here are effectively the icecast nsis files moved and made
more like a template. The names match for mingw but similar styles could be done
for other compilers. They only get used in the mingw case as determined by

The conditional compile is done a different way in automake, this seem to be
causing a pic/non-pic builds for those. so worked around that.