log contents not expanding timestamps
2 changes here really.  pull out the timestamp expansion code into a separate
routine so it can be called by both the log thread and the log_contents routine

Allow for the log content routine to take a level setting so that a selectable
set or messages can be requested. supply a level param to admin to supply that
else it defaults to whatever the log level is set to.

The only concern is that debug messages can fill up the history fairly quickly
so you do not get much.   We could use the entries to refer to non-debug
messages instead but does that make it confusing.

karl.