log update
This is a split of the locking, a RW lock to replace the current mutex to make
sure the high level log info is static while log lines are being formed.

a rotating pool of log line entires on a freed list being reused to keep things
more isolated from allocations elsewhere.  We could pre-allocate a pool to set
things off, but for now it allocates a frees only if there is excessive amounts.

The mutex is pushed down to per-log queues so the scope of holding the lock is
reduced to the incoming and freed queues.

To help in core analysis, the logs with error log have a reserved space at the
beginning for a sequence number. This does not appear in the log but allows for
a quick analysis of the log entries for crash analysis whether they are written
or not.