fix a couple of issues on log re-opening
The stat call was not working well enough on windows, it seems to be lagging in
the size values (may of been specific to certain setups). With fstat it seems to
be working better, more in-line with the written data.

Fixed the log cycling on windows, and alignment for the EOL differences.

karl